moved geometry transport implementation to own directory
