+++ include/nektar_interface/basis_evaluation.hpp
@@ -11,7 +11,7 @@
 #include <StdRegions/StdExpansion2D.h>
 
 #include "function_coupling_base.hpp"
-#include "geometry_transport_3d.hpp"
+#include "geometry_transport/shape_mapping.hpp"
 #include "special_functions.hpp"
 #include "utility_sycl.hpp"
