+++ include/nektar_interface/geometry_transport/geometry_transport_2d.hpp
@@ -10,6 +10,8 @@
 #include <mpi.h>
 #include <vector>
 
+#include "shape_mapping.hpp"
+
 using namespace std;
 using namespace Nektar;
 using namespace Nektar::SpatialDomains;
