+++ include/nektar_interface/geometry_transport/geometry_transport_3d.hpp
@@ -14,6 +14,7 @@ using namespace Nektar::SpatialDomains;
 using namespace Nektar::LibUtilities;
 
 #include "geometry_transport_2d.hpp"
+#include "shape_mapping.hpp"
 
 namespace NESO {
 
@@ -252,28 +253,6 @@ inline void assemble_geometry_container_3d(
   for (auto geom : remote_geoms_3d) {
     output_container.push_back(geom);
   }
-}
-
-/**
- * Get the unqiue integer (cast) that corresponds to a Nektar++ shape type.
- *
- * @param shape_type ShapeType Enum value.
- * @returns Static cast of enum value.
- */
-inline constexpr int
-shape_type_to_int(Nektar::LibUtilities::ShapeType shape_type) {
-  return static_cast<int>(shape_type);
-}
-
-/**
- * Get the unqiue Enum (cast) that corresponds to an integer returned form
- * shape_type_to_int.
- *
- * @param type_int Int returned from shape_type_to_int.
- * @returns Static cast to Enum value.
- */
-inline Nektar::LibUtilities::ShapeType int_to_shape_type(const int type_int) {
-  return static_cast<Nektar::LibUtilities::ShapeType>(type_int);
 }
 
 /**
