+++ include/nektar_interface/particle_cell_mapping/particle_cell_mapping_2d.hpp
@@ -7,10 +7,10 @@
 #include <mpi.h>
 #include <vector>
 
-#include "../coordinate_mapping.hpp"
-#include "../geometry_transport_2d.hpp"
-#include "../particle_mesh_interface.hpp"
 #include "coarse_lookup_map.hpp"
+#include "nektar_interface/coordinate_mapping.hpp"
+#include "nektar_interface/geometry_transport/shape_mapping.hpp"
+#include "nektar_interface/particle_mesh_interface.hpp"
 #include "newton_geom_interfaces.hpp"
 #include "particle_cell_mapping_common.hpp"
