+++ include/nektar_interface/particle_mesh_interface.hpp
@@ -19,8 +19,7 @@
 #include <neso_particles.hpp>
 
 #include "bounding_box_intersection.hpp"
-#include "geometry_transport_2d.hpp"
-#include "geometry_transport_3d.hpp"
+#include "geometry_transport/geometry_transport.hpp"
 #include "particle_boundary_conditions.hpp"
 
 using namespace Nektar::SpatialDomains;
