+++ solvers/Electrostatic2D3V/ParticleSystems/charged_particles.hpp
@@ -3,7 +3,7 @@
 
 #include <nektar_interface/function_evaluation.hpp>
 #include <nektar_interface/function_projection.hpp>
-#include <nektar_interface/halo_extension.hpp>
+#include <nektar_interface/geometry_transport/halo_extension.hpp>
 #include <nektar_interface/particle_interface.hpp>
 
 #include <neso_particles.hpp>
