+++ test/unit/nektar_interface/test_particle_geometry_interface.cpp
@@ -1,4 +1,4 @@
-#include "nektar_interface/halo_extension.hpp"
+#include "nektar_interface/geometry_transport/halo_extension.hpp"
 #include "nektar_interface/particle_interface.hpp"
 #include "nektar_interface/utility_mesh_plotting.hpp"
 #include <LibUtilities/BasicUtils/SessionReader.h>
@@ -463,7 +463,6 @@ TEST(ParticleGeometryInterface, PBC) {
 
 TEST(ParticleGeometryInterface, HaloExtend2D) {
   const int width = 1;
-
 
   LibUtilities::SessionReaderSharedPtr session;
   SpatialDomains::MeshGraphSharedPtr graph;
